scripts/test_api_key_priority.py
#!/usr/bin/env python3
"""
Test script to verify API key priority handling.
"""

import asyncio
import os
import sys
from pathlib import Path

# Add parent directory to path
sys.path.insert(0, str(Path(__file__).parent.parent))

from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Test with the MultiModelService directly
async def test_api_key_priority():
    from app.multi_model_service import MultiModelService
    from app.models import ModelProvider
    
    service = MultiModelService()
    
    print("Testing API Key Priority Handling\n")
    
    # Test 1: User key should be used when provided
    print("Test 1: Using user-provided API key")
    try:
        # This should fail because it contains a placeholder pattern
        # Check if placeholder detection is working
        test_key = "your_test_key_here"
        is_placeholder = service._is_placeholder_key(test_key)
        print(f"  Is '{test_key}' detected as placeholder? {is_placeholder}")
        
        model = service._get_model(ModelProvider.GEMINI, is_evaluator=False, api_key=test_key)
        print("❌ FAILED: Should have rejected placeholder key")
    except ValueError as e:
        print(f"✅ PASSED: {e}")
    except Exception as e:
        print(f"✅ PASSED (API call failed): {type(e).__name__}: {e}")
    
    # Test 2: Server key should be used when no user key provided
    print("\nTest 2: Using server API key (fallback)")
    try:
        model = service._get_model(ModelProvider.GEMINI, is_evaluator=False, api_key=None)
        print("❌ FAILED: Should have rejected server placeholder key")
    except ValueError as e:
        print(f"✅ PASSED: {e}")
    
    # Test 3: Valid user key format (simulated)
    print("\nTest 3: Valid API key format")
    try:
        # Simulate a valid-looking key
        valid_key = "AIzaSyB1234567890abcdefghijklmnopqrstuv"
        model = service._get_model(ModelProvider.GEMINI, is_evaluator=False, api_key=valid_key)
        print("✅ PASSED: Accepted valid-format key")
    except Exception as e:
        print(f"✅ PASSED: Key format accepted, actual API call would fail: {e}")
    
    # Test 4: Evaluator should also use user key first
    print("\nTest 4: Evaluator uses user key priority")
    try:
        # This should fail because 'your_api_key_here' is a placeholder
        model = service._get_model(ModelProvider.GEMINI, is_evaluator=True, api_key="your_api_key_here")
        print("❌ FAILED: Should have rejected placeholder key")
    except ValueError as e:
        print(f"✅ PASSED: {e}")

if __name__ == "__main__":
    asyncio.run(test_api_key_priority())