requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
langchain
langchain-google-genai
langchain-openai
langchain-anthropic
python-dotenv
structlog
python-multipart
httpx
requests
pytest
pytest-asyncio
bleach
sqlalchemy[asyncio]
asyncpg
aiosqlite
alembic
python-jose[cryptography]
passlib[bcrypt]
cryptography
pydantic[email]